pyproject.toml
[project]
name = "dagster-poc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dagster>=1.10.14",
    "dagster-duckdb>=0.26.14",
    "dagster-webserver>=1.10.14",
    "pandas>=2.2.3",
    "pydantic>=2.11.4",
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
    "ruff>=0.11.9",
]
